day_13/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
python-jose[cryptography]
pydantic
pydantic-settings
email-validator